user_app/requirements.txt
uvicorn==0.21.1
fastapi==0.95.0
pymongo==4.3.3
python-jose==3.3.0
passlib==1.7.4
motor==3.1.2
fastapi_jwt_auth==0.5.0
python-dotenv==1.0.0